+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Fix bug in RandomCrop and segmentation dataset getitem by `@illian01` in [PR 535](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/535)
 - Handle unexpected error during training rt-detr by `@hglee98` in [PR 530](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/530)
+- Fix no 'deploy' attribute error in fx model train phase by `@illian01` in [PR 543](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/543)
 
 ## Breaking Changes: